+++ components/molecules/searchForm/index.tsx
@@ -3,23 +3,14 @@
 /* eslint-disable no-use-before-define */
 import SearchBtn from 'components/atoms/searchBtn';
 import SearchInput from 'components/atoms/searchInput';
-import useReduxData from 'hooks/useReduxData';
+import useFilterByQueryParam from 'hooks/useFilterByQueryParam';
 import React, { ChangeEvent, FormEvent, useState } from 'react';
-import { useAppDispatch } from 'redux/configStore';
-import { setFilteredPostList } from 'redux/modules/post';
 import { appendQueryString } from 'util/common';
-import FilterByQueryParam from 'util/filterByQueryParam';
 import styles from './index.module.scss';
 
 const SearchForm = () => {
-  const { originalNotionList } = useReduxData();
-  const dispatch = useAppDispatch();
+  const { handleFilter } = useFilterByQueryParam();
   const [searchValue, setSearchValue] = useState<string>('');
-
-  const handleFilter = () => {
-    const filter = new FilterByQueryParam(originalNotionList);
-    dispatch(setFilteredPostList(filter.filteredList()));
-  };
 
   const handleFormSubmit = () => (e: FormEvent<HTMLFormElement>) => {
     e.preventDefault();
